% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsifyr-package.R
\docType{package}
\name{falsifyr-package}
\alias{falsifyr}
\alias{falsifyr-package}
\title{falsifyr: adversarial robustness checks for statistical claims}
\description{
falsifyr attacks fitted model claims with row deletion, alternative standard
errors, covariate drops, missing-data perturbations, measurement error,
placebo checks, and bounded specification search.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/msaule/falsifyr}
  \item \url{https://msaule.github.io/falsifyr/}
  \item Report bugs at \url{https://github.com/msaule/falsifyr/issues}
}

}
\author{
\strong{Maintainer}: Markuss Saule \email{markusstomas@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Markuss Saule \email{markusstomas@gmail.com} [copyright holder]
}

}
\keyword{internal}
